across line link and move text area into own component
