+++ src/styles/TextViewer.module.css
@@ -1,10 +1,5 @@
-.text {
+.text_viewer {
   font-size: 13px;
-}
-
-.text_viewer_container {
-  margin: 0 auto;
-  background: white;
 }
 
 .layout_header {
@@ -17,15 +12,6 @@
   display: flex;
 }
 
-.main_container {
-  padding: 16px;
-  border: 1px solid #ccc;
-  border-top: none;
-  min-height: calc(100vh - 50px);
-  overflow-y: auto;
-  flex: 1;
-}
-
 .metadata_side_container {
   flex: 0 0 250px;
   padding: 8px 0;
@@ -34,8 +20,4 @@
 .attributes_side_container {
   flex: 0 0 250px;
   padding: 8px;
-}
-
-.text_container {
-  line-height: 40px;
 }
